make new style acceptance test the default
